+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Recent and upcoming changes to lightdash
 
 ### Fixed
 - Fixed a bug where chart config would reset after running a query
+- Fixed a bug in the chart where we could only select 1 metric when we were grouping by dimension
 
 ## [0.8.1] - 2021-09-29
 ### Added